rg: don't embellish performance
ripgrep is fast, but no non-toy search tool is going to be faster in
every case. It's better to be a bit more reserved instead of spreading
misinformation about the performance of a tool. It may be an obvious
embellishment, but plenty of people will take it seriously.

In general, I think we should encourage spreading the nuance in the
trade offs of our tools as opposed to hyping them.